+++ gif.h
@@ -16,8 +16,8 @@
 extern int GIF_OpenFile(const char *,const char *, unsigned int, unsigned int);
 extern void GIF_CloseFile(void);
 extern unsigned int GIF_GetPixel(int,int,int);
-extern int GIF_GetWidth(void);
-extern int GIF_GetHeight(void);
+extern size_t GIF_GetWidth(void);
+extern size_t GIF_GetHeight(void);
 
 #endif
