+++ png.h
@@ -7,8 +7,8 @@
 extern int PNG_OpenFile(const char *,const char *, unsigned int, unsigned int);
 extern void PNG_CloseFile(void);
 extern unsigned int PNG_GetPixel(int,int,int);
-extern int PNG_GetWidth(void);
-extern int PNG_GetHeight(void);
+extern size_t PNG_GetWidth(void);
+extern size_t PNG_GetHeight(void);
 
 #endif
